Fix numUnSigned and wrong byte cast
The connector was incorrectly only supplying byte values to Polytune via
the `numUnsigned` method. In the case of the length of UUID values,
it was casted to byte resuling in an overflow for large arrays.

This changes the numUnsigned method to take uint instead, which is
a superset of the integer values that need to be supplied to Polytune.
If a too large value is supplied with a dataType `u8`, this will
be rejected by Polytune.